三方库初始化优化
